+++ src/app/Categorias/Producto/contador/styles.css
@@ -1,0 +1,45 @@
+.count-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.count-container__contador {
+  display: grid;
+  grid-template: 1fr / 1fr 1fr 1fr;
+  width: 140px;
+  height: 50px;
+  font-weight: 600;
+  font-size: 1.4rem;
+  margin: 40px auto 0;
+}
+
+.count-container__button {
+  color: var(--primary-color);
+  width: 100%;
+  height: 100%;
+  border: 1px solid var(--border-color);
+  font-size: 1.6rem;
+  transition: all 250ms ease-out;
+}
+
+.count-container__button:hover {
+  border: none;
+  color: var(--just-white);
+  background-color: var(--primary-color);
+}
+
+.count-container__button[disabled] {
+  color: var(--border-color);
+  background-color: transparent;
+  border: 1px solid var(--border-color);
+}
+
+.count-container__qty {
+  text-align: center;
+  height: 100%;
+  padding-top: 10px;
+  color: var(--primary-color);
+  background-color: var(--just-white);
+}
